REFAC: rename internal library
(cherry picked from commit 409c286a057f48000bb40f0ada246e79f3ece400)
